Domasna 4/Microservices/lstm-prediction-service/requirements.txt
fastapi
uvicorn
requests
cachetools

numpy<2.0
pandas
scikit-learn

tensorflow-cpu
psycopg2-binary
python-dotenv

